Refactor app code
