+++ pages/[version]/[...slug].tsx
@@ -1,12 +1,8 @@
 import Markdoc from "@markdoc/markdoc";
 import { isEqual, isObject, isUndefined } from "lodash";
 import { GetServerSidePropsContext, GetServerSidePropsResult } from "next";
-import Head from "next/head";
-import Script from "next/script";
-import React, { useMemo } from "react";
-import ErrorBoundary from "../../components/ErrorBoundary";
+import { useMemo } from "react";
 import Footer from "../../components/Footer/Footer";
-import GoogleAnalyticsScript from "../../components/GoogleAnalyticsScript/GoogleAnalyticsScript";
 import APIPageLayout from "../../components/PageLayouts/APIPageLayout/APIPageLayout";
 import DocsPageLayout from "../../components/PageLayouts/DocsPageLayout/DocsPageLayout";
 import { SelectOption } from "../../components/SelectDropdown/SelectDropdown";
@@ -71,53 +67,18 @@ export default function Article({
       : { value: false, pageInfoObject: { label: "", value: "" } };
   }, [slug]);
 
-  return (
-    <>
-      <Head>
-        <title>{pageTitle}</title>
-        <meta content={pageTitle} property="og:title" />
-        <meta content={pageTitle} name="twitter:title" />
-        {pageDescription && (
-          <React.Fragment>
-            <meta content={pageDescription} name="description" />
-            <meta content={pageDescription} property="og:description" />
-            <meta content={pageDescription} name="twitter:description" />
-          </React.Fragment>
-        )}
-        <Script
-          id="show-banner"
-          dangerouslySetInnerHTML={{
-            __html: `              
-                (function(window, document, dataLayerName, id) {
-                  window[dataLayerName]=window[dataLayerName]||[],window[dataLayerName].push({start:(new Date).getTime(),event:"stg.start"});var scripts=document.getElementsByTagName('script')[0],tags=document.createElement('script');
-                  function stgCreateCookie(a,b,c){var d="";if(c){var e=new Date;e.setTime(e.getTime()+24*c*60*60*1e3),d="; expires="+e.toUTCString()}document.cookie=a+"="+b+d+"; path=/"}
-                  var isStgDebug=(window.location.href.match("stg_debug")||document.cookie.match("stg_debug"))&&!window.location.href.match("stg_disable_debug");stgCreateCookie("stg_debug",isStgDebug?1:"",isStgDebug?14:-1);
-                  var qP=[];dataLayerName!=="dataLayer"&&qP.push("data_layer_name="+dataLayerName),isStgDebug&&qP.push("stg_debug");var qPString=qP.length>0?("?"+qP.join("&")):"";
-                  tags.async=!0,tags.src="https://collate.containers.piwik.pro/"+id+".js"+qPString,scripts.parentNode.insertBefore(tags,scripts);
-                  !function(a,n,i){a[n]=a[n]||{};for(var c=0;c<i.length;c++)!function(i){a[n][i]=a[n][i]||{},a[n][i].api=a[n][i].api||function(){var a=[].slice.call(arguments,0);"string"==typeof a[0]&&window[dataLayerName].push({event:n+"."+i+":"+a[0],parameters:[].slice.call(arguments,1)})}}(i[c])}(window,"ppms",["tm","cm"]);
-                  })(window, document, 'dataLayer', '85b94982-8c42-497f-96c9-353365f1fe7a');
-                `,
-          }}
-        />
-        <GoogleAnalyticsScript />
-      </Head>
-
-      <ErrorBoundary>
-        {isAPIsPage.value ? (
-          <APIPageLayout
-            parsedContent={parsedContent}
-            pageInfoObject={isAPIsPage.pageInfoObject}
-          />
-        ) : (
-          <DocsPageLayout
-            navbar={<TopNav versionsList={versionsList} />}
-            parsedContent={parsedContent}
-            slug={slug}
-            footer={<Footer />}
-          />
-        )}
-      </ErrorBoundary>
-    </>
+  return isAPIsPage.value ? (
+    <APIPageLayout
+      parsedContent={parsedContent}
+      pageInfoObject={isAPIsPage.pageInfoObject}
+    />
+  ) : (
+    <DocsPageLayout
+      navbar={<TopNav versionsList={versionsList} />}
+      parsedContent={parsedContent}
+      slug={slug}
+      footer={<Footer />}
+    />
   );
 }
