+++ pages/[version]/index.tsx
@@ -1,9 +1,7 @@
-import Head from "next/head";
-import React, { useEffect } from "react";
+import { useEffect } from "react";
 import CategoriesNav from "../../components/CategoriesNav/CategoriesNav";
 import ConnectorsInfo from "../../components/ConnectorsInfo/ConnectorsInfo";
 import Footer from "../../components/Footer/Footer";
-import GoogleAnalyticsScript from "../../components/GoogleAnalyticsScript/GoogleAnalyticsScript";
 import NewsEntry from "../../components/NewsEntry/NewsEntry";
 import { SelectOption } from "../../components/SelectDropdown/SelectDropdown";
 import TopNav from "../../components/TopNav/TopNav";
@@ -23,10 +21,6 @@ import { useRouteChangingContext } from "../../context/RouteChangingContext";
 import { SkeletonWidth } from "../../enums/SkeletonLoder.enum";
 import { getVersionsList } from "../../lib/api";
 
-const TITLE = "OpenMetadata Documentation: Get Help Instantly";
-const DESCRIPTION =
-  "Follow the step-by-step guides to get started with OpenMetadata, the #1 open source data catalog tool. Get discovery, collaboration, governance, observability, quality tools all in one place.";
-
 interface Props {
   versionsList: Array<SelectOption<string>>;
 }
@@ -44,19 +38,6 @@ export default function Index({ versionsList }: Readonly<Props>) {
 
   return (
     <>
-      <Head>
-        <title>{TITLE}</title>
-        <meta content={TITLE} property="og:title" />
-        <meta content={TITLE} name="twitter:title" />
-        {DESCRIPTION && (
-          <React.Fragment>
-            <meta content={DESCRIPTION} name="description" />
-            <meta content={DESCRIPTION} property="og:description" />
-            <meta content={DESCRIPTION} name="twitter:description" />
-          </React.Fragment>
-        )}
-        <GoogleAnalyticsScript />
-      </Head>
       <div className="nav-bar-container">
         <TopNav versionsList={versionsList} />
         <CategoriesNav menu={menuItems} />
